new code structure, added (maybe works) sma on tmtng60
