+++ keyboards/pcilabs/pcilabs_common/multiplexer/multiplexer.h
@@ -5,11 +5,6 @@ SPDX-License-Identifier: GPL-2.0-or-later
 */
 #include <stdint.h>
 #include "quantum.h"
-#ifndef MUXES
-#    define MUXES 0
-#endif
-#if MUXES != 0
-
 uint8_t current_channel;
 
 void multiplexer_init(void);
@@ -26,4 +21,3 @@ extern const pin_t mux_selector_pins[MUX_SELECTOR_BITS];
 extern const mux_t mux_index[MUXES][MUX_CHANNELS];
 extern const mux_t NC;
 
-#endif
