+++ keyboards/pcilabs/pcilabs_common/multiplexer/multiplexer.mk
@@ -1,0 +1,5 @@
+SRC += \
+	pcilabs_common/multiplexer/multiplexer.c \
+	pcilabs_common/multiplexer/matrix_mux.c \
+	pcilabs_common/multiplexer/sma_mux.c
+
