+++ keyboards/pcilabs/pcilabs_common/pcilabs_common.mk
@@ -1,0 +1,8 @@
+SRC += \
+	analog.c \
+	pcilabs_common/lut.c \
+	pcilabs_common/sma.c \
+	pcilabs_common/scanfunctions.c \
+	pcilabs_common/custom_analog.c \
+	pcilabs_common/pcilabs_common.c \
+	pcilabs_common/matrix_analog.c
