+++ keyboards/pcilabs/pcilabs_common/scanfunctions.h
@@ -22,8 +22,6 @@ void matrix_read_cols_continuous_dynamic_actuation(matrix_row_t *current_row, ui
 
 void matrix_read_cols_static_actuation(matrix_row_t *current_row, uint8_t current_col, analog_key_t *key);
 
-void SMA_init(void);
-
-void SMA_filter(analog_key_t *key);
-
 void lut_init(void);
+
+extern void (*matrix_read_mode_array[7])(matrix_row_t*, uint8_t, analog_key_t*); // 3 bits for modes
