+++ keyboards/pcilabs/pcilabs_common/sma.h
@@ -1,0 +1,10 @@
+/*
+Copyright 2024 Mkass420 (@Mkass420)
+SPDX-License-Identifier: GPL-2.0-or-later
+*/
+#pragma once
+
+#include "analogkeys.h"
+
+void SMA_init(void);
+void SMA_filter(analog_key_t *key);
