+++ keyboards/pcilabs/tmtng60_moon/config.h
@@ -3,7 +3,7 @@ Copyright 2024 Mkass420 (@Mkass420)
 SPDX-License-Identifier: GPL-2.0-or-later
 */
 #pragma once
-#include "pcilabs_config_common.h"
+#include "pcilabs_common/pcilabs_config_common.h"
 
 #define RAW_USAGE_PAGE 0xFF60
 #define RAW_USAGE_ID 0x61
