+++ keyboards/pcilabs/tmtng60_moon/rules.mk
@@ -1,23 +1,23 @@
-#QUANTUM_LIB_SRC += analog.c
 MCU_FAMILY = STM32
 MCU_SERIES = STM32F1xx
-
 BOARD = STM32_F103_STM32DUINO
-
-SRC += matrix.c lut.c scanfunctions.c multiplexer.c custom_analog.c analog.c pcilabs_common.c matrix_common.c
 
 MCU_LDSCRIPT = GD32F303xC-dapboot
 BOOTLOADER = custom
-SRC += dapboot/bootloader.c
+SRC += keyboards/pcilabs/pcilabs_common/dapboot/bootloader.c
+
+include keyboards/pcilabs/pcilabs_common/pcilabs_common.mk
+include keyboards/pcilabs/pcilabs_common/multiplexer/multiplexer.mk
 
 RAW_ENABLE = yes
 WAIT_FOR_USB = yes
 NO_USB_STARTUP_CHECK = no
 
-CUSTOM_MATRIX = lite
-
-OPT = 2
-
 WEAR_LEVELING_DRIVER = embedded_flash
 
+OPT = 2
 CFLAGS += -fplan9-extensions
+
+CUSTOM_MATRIX = lite
+SRC += matrix.c
+
